utworzono w serisie nowy package z comparatorami
