+++ TripPlaner/src/main/java/com/singapore/TripPlaner/Controller/PlaceController.java
@@ -1,15 +1,12 @@
 package com.singapore.TripPlaner.Controller;
 
-import com.singapore.TripPlaner.Model.Places;
 import com.singapore.TripPlaner.Service.PlaceService;
-import com.singapore.TripPlaner.Service.PlacesComparatorBiggestRate;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.Collections;
 import java.util.List;
 
 @Controller
