+++ TripPlaner/src/main/java/com/singapore/TripPlaner/Service/comparators/PlacesComparatorBiggestRate.java
@@ -1,16 +1,16 @@
-package com.singapore.TripPlaner.Service;
+package com.singapore.TripPlaner.Service.comparators;
 
 import com.singapore.TripPlaner.Model.Places;
 import org.springframework.stereotype.Component;
-import java.util.Comparator;
 
+import java.util.Comparator;
 @Component
-public class PlacesComparatorMostPopular implements Comparator<Places> {
+public class PlacesComparatorBiggestRate implements Comparator<Places> {
     @Override
     public int compare(Places o1, Places o2) {
-        int compareResult = Integer.compare(o2.getNumberOfOpinions(),o1.getNumberOfOpinions());
+        int compareResult = Double.compare(o2.getRate(), o1.getRate());
         if (compareResult == 0) {
-            compareResult = Double.compare(o2.getRate(), o1.getRate());
+            compareResult = Integer.compare(o2.getNumberOfOpinions(),o1.getNumberOfOpinions());
         }
         if (compareResult == 0) {
             compareResult = o1.getName().compareToIgnoreCase(o2.getName());
@@ -18,4 +18,3 @@ public class PlacesComparatorMostPopular implements Comparator<Places> {
         return compareResult;
     }
 }
-
